Snappy pants snaps parts together
just use matching parts and colliders and put is trigger and snappyPants on the dominant part. The dominant part will receive the snap on in perfect position on trigger from the colliders and then childs the recessive part (one without the script) to the dominant part which presumably you are dragging twith mouse script or VR controllers.